Simplify DelegatingSerializationService & expose enterprise serialization service registration hook
