+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/execution/ExecutionSerializerHooks.java
@@ -50,10 +50,6 @@ class ExecutionSerializerHooks {
                 }
 
                 @Override
-                public void destroy() {
-                }
-
-                @Override
                 public void write(ObjectDataOutput out, DoneItem object) {
                 }
 
@@ -83,10 +79,6 @@ class ExecutionSerializerHooks {
                 @Override
                 public int getTypeId() {
                     return SerializerHookConstants.SNAPSHOT_BARRIER;
-                }
-
-                @Override
-                public void destroy() {
                 }
 
                 @Override
@@ -124,11 +116,6 @@ class ExecutionSerializerHooks {
                 }
 
                 @Override
-                public void destroy() {
-
-                }
-
-                @Override
                 public void write(ObjectDataOutput out, BroadcastEntry object) throws IOException {
                     out.writeObject(object.getKey());
                     out.writeObject(object.getValue());
@@ -160,11 +147,6 @@ class ExecutionSerializerHooks {
                 @Override
                 public int getTypeId() {
                     return SerializerHookConstants.BROADCAST_KEY;
-                }
-
-                @Override
-                public void destroy() {
-
                 }
 
                 @Override
