+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/execution/init/CustomClassLoadedObject.java
@@ -117,10 +117,5 @@ public final class CustomClassLoadedObject {
         public int getTypeId() {
             return SerializerHookConstants.CUSTOM_CLASS_LOADED_OBJECT;
         }
-
-        @Override
-        public void destroy() {
-
-        }
     }
 }
