+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/config/JobConfigTest.java
@@ -147,9 +147,5 @@ public class JobConfigTest extends JetTestSupport {
         public Object read(ObjectDataInput in) {
             return null;
         }
-
-        @Override
-        public void destroy() {
-        }
     }
 }
